Convert a .Rmd file into a .html file according to finddx design / template.

Convert a .Rmd file into a .html file according to finddx design / template.

html_document_find(
  toc = TRUE,
  toc_depth = 4,
  toc_float = TRUE,
  code_folding = "show",
  ...
)

Arguments

toc

Boolean. To include a table of contents in the output. Default is TRUE.

toc_depth

Numeric. Depth of headers to include in table of contents. Default is 4.

toc_float

Boolean. To float the table of contents to the left of the document. Default is TRUE.

code_folding

Character. Enable document readers to toggle the display of R code chunks.

...

Additional arguments of rmarkdown::html_document().

Value

An R Markdown output format object

Examples

# Store in temp file for the example
rmddir <- tempfile(pattern = "rmd-")
dir.create(rmddir)

# Knit
rmarkdown::render(input = system.file("rmarkdown", "rmd_template.Rmd", package = "finddxtemplate"), 
                  output_format = html_document_find(),
                  output_dir = rmddir)
#> 
#> 
#> processing file: rmd_template.Rmd
#>   |                                                                              |                                                                      |   0%  |                                                                              |.....                                                                 |   7%
#>    inline R code fragments
#> 
#>   |                                                                              |..........                                                            |  14%
#> label: unnamed-chunk-1 (with options) 
#> List of 1
#>  $ include: logi FALSE
#> 
#>   |                                                                              |...............                                                       |  21%
#>   ordinary text without R code
#> 
#>   |                                                                              |....................                                                  |  29%
#> label: unnamed-chunk-2 (with options) 
#> List of 1
#>  $ echo: logi FALSE
#> 
#>   |                                                                              |.........................                                             |  36%
#>   ordinary text without R code
#> 
#>   |                                                                              |..............................                                        |  43%
#> label: unnamed-chunk-3
#>   |                                                                              |...................................                                   |  50%
#>   ordinary text without R code
#> 
#>   |                                                                              |........................................                              |  57%
#> label: unnamed-chunk-4 (with options) 
#> List of 1
#>  $ echo: logi FALSE
#> 
#>   |                                                                              |.............................................                         |  64%
#>   ordinary text without R code
#> 
#>   |                                                                              |..................................................                    |  71%
#> label: unnamed-chunk-5 (with options) 
#> List of 3
#>  $ fig.width : num 10
#>  $ fig.height: num 6
#>  $ fig.align : chr "center"
#> 
#>   |                                                                              |.......................................................               |  79%
#>   ordinary text without R code
#> 
#>   |                                                                              |............................................................          |  86%
#> label: unnamed-chunk-6
#>   |                                                                              |.................................................................     |  93%
#>   ordinary text without R code
#> 
#>   |                                                                              |......................................................................| 100%
#> label: unnamed-chunk-7 (with options) 
#> List of 3
#>  $ fig.width : num 10
#>  $ fig.height: num 6
#>  $ fig.align : chr "center"
#> 
#> 
#> output file: rmd_template.knit.md
#> /usr/bin/pandoc +RTS -K512m -RTS rmd_template.knit.md --to html4 --from markdown+autolink_bare_uris+tex_math_single_backslash --output /tmp/Rtmpjrcw19/rmd-32c75e74ea7a/rmd_template.html --lua-filter /home/runner/work/_temp/Library/rmarkdown/rmarkdown/lua/pagebreak.lua --lua-filter /home/runner/work/_temp/Library/rmarkdown/rmarkdown/lua/latex-div.lua --self-contained --variable bs3=TRUE --standalone --section-divs --table-of-contents --toc-depth 4 --variable toc_float=1 --variable toc_selectors=h1,h2,h3,h4 --variable toc_collapsed=1 --variable toc_smooth_scroll=1 --variable toc_print=1 --template /home/runner/work/_temp/Library/finddxtemplate/html_template.html --highlight-style pygments --variable theme=bootstrap --css /home/runner/work/_temp/Library/finddxtemplate/css/main.css --include-in-header /tmp/Rtmpjrcw19/rmarkdown-str32c7117ec48b.html --variable code_folding=show --variable code_menu=1 
#> 
#> Output created: /tmp/Rtmpjrcw19/rmd-32c75e74ea7a/rmd_template.html

# Open the knitted example
browseURL(file.path(rmddir, "rmd_template.html"))